requirements.txt
anthropic
gradio>=4.0
pdfplumber
python-dotenv
